requirements.txt
fastapi
uvicorn
streamlit
supabase
twilio
python-dotenv
stripe
openai
python-multipart
werkzeug
